[explorer] use suix_getEpochMetrics to fetch home page data to speed up performance (#14865)
## Description 

We use `suix_getEpochs` to fetch metrics for the transaction block graph
and epochs tab, but this API returns validator data for each epoch which
ends up transmitting a lot of data (~6MB) to the client. This also
results in the API taking ~5 seconds end-to-end.

@gegaowp made a new `suix_getEpochMetrics` that is similar to the above
endpoint but leaves out the validator data, so we can use that instead +
now the page is a lot snappier 😄

Before:
https://suiexplorer.com/

After:
https://explorer-git-wrobertson-exploerspeed-mysten-labs.vercel.app/

## Test Plan 
- Manual testing (chart data matches prod, table loads and paginates
correctly)

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes