[move-compiler] Coin field linter fix (#15096)
## Description 

This fixes a coin-field linter bug - Coin was identified by module name
only rather than by module name AND struct name...

## Test Plan 

Added a test checking that the spurious linter warning message is not
generated.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Coin-field linter will no longer generate a spurious warning message
when a field of a given object is from the `coin` module but is not a
`Coin` itself.