[move-ide] Bundling of move-analyzer binary with the plugin (#15060)
## Description 

This PR adds bundling of platform-specific move-analyzer binary with the
VSCode extension distribution itself. This seems like the current
solution of choice (e.g., adopted by newer versions of rust-analyzer).

The PR includes required changes to TS code as well as a script
responsible for packaging bundled extensions locally (for testing) as
well as publishing them.

Added licensing information copied from the root of the Sui repository
to have it included in VSCode Marketplace once the extension is
published.

This PR is marked as not having user-visible impact as it's not a change
to the existing extension but rather one of the initial steps to
creating a new one.

## Test Plan 

I tested this pretty extensively manually on a Mac (not sure if it's
even possible to automate this). I also tested bundled extension
installation (as well as manual installation described in the
Troubleshooting section) on Linux and Windows.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes