Return more accurate error for missing input object (#15114)
## Description 

When an object is missing from the store, instead of always return
ObjectNotFound, we try to be more smart about the error. This is
important because different validators could have different view of an
old object due to pruning. So when queried using an old object, we want
validators to return a more consistent error result.

## Test Plan 

TODO: I will add some tests.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes