[data ingestion] mvp for checkpoint consumer daemon (#15101)
initial minimal version of data ingestion daemon

### Structure 
The `IndexerExecutor` drives the process, employing the following
components:
* reader to consume new `CheckpointData` (currently, only `LocalReader`
is available, subscribing to changes in the local directory through the
`inotify` API).
* progress store to track the progress of individual pipelines
* vector of `WorkerPool`, with each pool dedicated to a specific type of
task.
Example usage can be found in the `basic_flow` test

### Flow
The Executor reads the current set of watermarks from the progress store
and initializes worker pools and readers. The reader sends updates to
the executor, which, in turn, broadcasts tasks to the pools. Each pool
manages its own set of workers, with concurrency defined per pool and
adjustable based on specific task requirements.
Upon receiving new checkpoint, the pool determines the worker to utilize
and proxies the task accordingly. Tasks are processed in parallel and
can finish in a random order. Once a continuous sequence from the
previous known watermark is formed, the worker pool reports the progress
to the executor, which updates the progress store.
When all worker pools are done with a given checkpoint, the reader is
notified to clean the directory.

### Idempotence
Tasks are expected to be idempotent. For certain tasks, one approach to
achieving idempotence is to verify at the start of the task whether it
has already been processed. Alternatively, for some of the workflows, a
concurrency of 1 can be used. In such cases, idempotence is irrelevant

### Misc notes
PR contains basic implementation of DynamoDB-based progress store and a
workflow that uploads checkpoint files to a S3 bucket.