[GraphQL] Refactor Validator and remove all PartialEq and Eq from Epoch, Validator, ValidatorSet, SystemState (#15243)
## Description 

Refactor Validator and remove all PartialEq and Eq from Epoch,
Validator, ValidatorSet, SystemState. This is in preparation for adding
the `apy` field to the type, as it requires extra data from the indexer
which needs to be passed in a method; using TryFrom wouldn't work.

## Test Plan 

Existing tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes