[move-ide] Extended on-hover for struct types (#15308)
## Description 

This PR extends on-hover functionality for struct types - when hovering
over the type, instead of just seeing the type name, what is now
displayed is the whole structure of the type (including fields). In
addition to being more informative, this is also more in line with what
rust-analyzer provides.

## Test Plan 

Test have been adjusted to reflect this change.