[gql][fix] connections return correct values for last queries (#15431)
## Description 

The bug is that when only last is provided, because the current
implementation depends on using before or after to determine how to
order the db, the code ends up returning the behavior as if we provided
first. To address this, rather than pass first and last directly to
pg_backend, I created a PageLimit enum with variants First, Last.
db_data_provider is responsible for determining whether to set the
variant as First or Last, and defaults to First(default_page_size) if a
value is not provided. Then, pg_backend maps First -> asc and Last ->
desc, optionally adding a filter for First -> after and Last -> before

## Test Plan 

Existing tests should work. Add new tests for connections with just
(last: n) to snapshot expected results.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes