[sui-node/metrics] recreate client (#15430)
## Description 
In certain cases where we experience degraded metric submit rates or an
inability to submit any, we can experience recovery issues when that
degredataion subsides. It is an edge case but it seems to present every
so often. The intent of this pr is to recreate a reqwest client anytime
we hit an error submitting metrics. This will allow for automatic
recovery of the node without restarting. Since our tick interval for
submit is every 60 secs, it should be fine to recreate.

## Test Plan 
local instruments

```
2023-12-19T20:31:02.903599Z  WARN sui_node::metrics: unable to push metrics: error sending request for url (https://metrics-proxy.testnet.sui.io:8443/publish/metrics): error trying to connect: tcp connect error: Connection refused (os error 61); new client will be created
```
and it recovers:

```
2023-12-19T20:31:18.018332Z  INFO sui_proxy::peers: 4 peers managed to make it on the allow list
[crates/sui-proxy/src/handlers.rs:54] "great success" = "great success"
2023-12-19T20:31:21.528435Z  INFO tower_http::trace::on_response: finished processing request latency=0.009472292 s status=201
```
---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes