[move] Reworked path parsing and extended move/copy usage  (#15267)
## Description 

- In the name of future proofing, reworked paths to work with `copy` and
`move` in addition to `&` and `&mut`. These new cases are feature gated
so we have time to make changes.
- Added a warning for seemingly borrowing a constant when it is actually
being copied. This warning is not feature gated.

## Test Plan 

- Added new tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Move will now warn on usage of constants that appear to borrow the value, but are actually borrowing a copy of the value.