[GraphQL/Cursor] Checkpoint querying and pagination (#15526)
## Description

Introduce a generic function for paginating DB queries, and use it for
checkpoint pagination. The generic algorithm fixes handling of
`has_previous_page` and supports more combinations of
`first`/`after`/`last`/`before` (only combining `first` and `last` is
not supported currently, and even this could be permitted as long as we
specify that both limits are individually less than the page size).

Adopting the new pagination algorithm for checkpoints also required
moving it to the new `QueryExecutor` interface, which had several
knock-on implications, and/or moved my through parts of the codebase
that needed clean-ups:

- `Epoch`, `ProtocolConfigs`, `Query.availableRange` and
`Query.chainIdentifier` were all also moved to the new query interface.
- Querying for the `chainIdentifier` now no longer over-fetches the
entire checkpoint just to get the digest -- it only queries for the
digest (although once we support DataLoaders we may want to undo this).
- The queries issued as part of creating the `ProtocolConfigs` type have
been batched into a single query, to avoid doing multiple sequential
round-trips to the database.
- `ProtocolConfigs` has adopted the representation where it is largely
represented by its native form.
- `StakedSui`'s computed fields were incorrectly returning `Result<_,
Error>` rather than just `Result<_>`. This means that errors produced in
the course of fetching these fields would not be correctly annotated
with an error code.
- The `Epoch` fields that are part of transactions have been made
nullable to account for transactions that are dry-run (and therefore
could mention arbitrary epochs).

## Test Plan

Updated E2E tests related to checkpoint pagination:

```
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration         \
  -- checkpoint_connection_pagination.move
```

(+ all the other E2E tests to cover the other refactorings)

## Stack
- #15467 
- #15470 
- #15471 
- #15472 
- #15473
- #15474 
- #15475 
- #15484 
- #15485
- #15519
- #15520
- #15521
- #15522
- #15523
- #15524
- #15525