[transfer-to-object] Turn on for mainnet (#15533)
## Description 

Turns on transfer-to-object in mainnet.

## Test Plan 

Tested and has been running in devnet and testnet for the past couple
months.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] protocol change
- [X] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Enables transfer-to-object functionality on mainnet.