[gql][2/n] output nodes estimation (#15497)
## Description 

Implements output nodes estimation. A node adds a cost of 1, unless it
is a connection field, where it would then have a cost of
default_page_size, or min(first, last) if so provided. This is
multiplicative - at the next child connection, that child's estimated
output will be multiplied with its parent's to yield the cumulative
result. A connection's pageInfo field does not multiplicatively increase the current count, as there is only one node for the entire connection.

## Test Plan 

1. check that we get this back in the headers when showUsage is provided
4. output_node_estimation.move

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes