[cli][alias][easy] Allow the dot character in aliases' names (#15608)
## Description 

Allow the dot character in aliases' names so that someone can associate
an address like this: `my.address.sui`

## Test Plan 

Added test for validating aliases.

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes
Relaxed the rule around Sui CLI aliases names. The character `.` is now
allowed to be in an alias for the keys in the Sui CLI.