[feat] introduce the core thread dispatcher (#15610)
## Description 

Introduces the core thread dispatcher to the codebase. Compared to the
Mysticeti codebase there are a few differences:
* a `handle` is returned from the `CoreThreadDispatcher` when created.
The `handle` is used to `stop` the `CoreThreadDispatcher`.
* the `CoreThreadDispatcher` is now cloneable and shareable (ex behind
an `Arc`) compared to the previous implementation . That will allow more
than one components accessing `Core` , in contrast to Mysticeti codebase
where the `net_sync` component was the only one allowed to access it.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and do not break anything, you can
skip the following section. Otherwise, please briefly describe what has
changed under the Release Notes section.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes